coordinates are int

+++ point.py
@@ -2,7 +2,7 @@ from vector import Vector
 
 
 class Point:
-    def __init__(self, x: float, y: float):
+    def __init__(self, x: int, y: int):
         self.x = x
         self.y = y
