+++ vector.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 
 class Vector:
-    def __init__(self, x: float, y: float):
+    def __init__(self, x: int, y: int):
         self.x = x
         self.y = y
 
@@ -11,7 +11,7 @@ class Vector:
         y = self.y + other.y
         return Vector(x, y)
 
-    def __mul__(self, other: Vector) -> float:
+    def __mul__(self, other: Vector) -> int:
         return self.x * other.x + self.y * other.y
 
     def __eq__(self, other: object) -> bool:
